API

DegreesOfFreedom.anim_plot — Method

anim_plot(βs, βlasso)

Compare the lasso solution from glmnet and the iterative ridge via an animated plot.

DegreesOfFreedom.cvlasso_vs_iter_ridge — Function

cvlasso_vs_iter_ridge(n, p)

Compare lasso with LOOCV and iterative ridge.

DegreesOfFreedom.demo_lasso — Function

demo_lasso(n, p, p1)

Demo for Lasso fitted by iterative ridge regressions.

DegreesOfFreedom.demo_lasso_df — Function

demo_lasso_df(n, p)

Demo for degrees of freedom of lasso via the iterative ridge regression.

DegreesOfFreedom.df_regtree — Method

df_regtree(; ps = [1, 5, 10], maxd = 4)

Experiment for degrees of freedom for regression trees with number of features ps and maximum depth maxd.

DegreesOfFreedom.gen_data — Function

gen_data(n, p, p1)

Generate simulation data:

X of size nxp

β of size p, where only the first p1 elements are signal.

y of size p: y = Xβ + ε

DegreesOfFreedom.iter_ridge — Method

iter_ridge(X::AbstractMatrix, y::AbstractVector, λ::Real)

Conduct iterative ridge regression for y on X with smoothness penalty parameter λ.

DegreesOfFreedom.run_experiment_lasso_vs_subset — Method

run_experiment_lasso_vs_subset()

Run the experiment for comparing the degrees of freedom of lasso and best subset, whose results are saved into the figure shown in the paper.

DegreesOfFreedom.run_experiment_splines — Method

run_experiment_splines()

Run the experiment of splines, whose results will be saved into a .sil file (can be later loaded via deserialize) and a .tex file (the table displayed in the paper).

DegreesOfFreedom.run_experiment_tree — Method

run_experiment_tree()

Run the experiment for regression tree, whose results are saved into a .sil file (can be later loaded via deserialize) and a .tex file (the table in the paper).

DegreesOfFreedom.save_plots — Method

save_plots(ps; output)

Save multi-images into a pdf file, if output is unspecified (default), the resulting file is /tmp/all.pdf. See also: https://github.com/szcf-weiya/Xfunc.jl/blob/master/src/plot.jl